Further implementation of cleaning procedure
+ There are now two ways to clean images using standard deviation of
isoenergetic rows: directly using the photon_events (clean_image_std),
or converting it into a matrix and than back to photon_events
(clean_image_std_matrix). There is less messing with the data in the
first option, but it seems to be a bit slower (by ~ 30%, it runs one
2048x2048 image in ~ 2.9 sec, versus ~ 2.2 sec with the matrix option).

+ setup.py was update to include the ‘nexusformat’ and ‘pillow’ as
requirements.